[DEV] SEOS-946/SEOS-798: Merge "router" into API layer
The API now only knows to talk to a CLIENT or LIB instance and config and context structs were aligned to match this structure.
Proxy objects carry a pointer to the respective instance they were created through; this allows to do the "routing" on the API level already and also improves the "locality" tracking, which does not need to be done on a lower leve.